+++ merged-packages/bitcoin-wallet-snap/integration-test/keyring-request.test.ts
@@ -324,6 +324,7 @@ describe('KeyringRequestHandler', () => {
         result: {
           psbt: expect.any(String), // non deterministic
           txid: expect.any(String),
+          canBeMalleable: false,
         },
       });
 
@@ -639,6 +640,7 @@ describe('KeyringRequestHandler', () => {
         pending: false,
         result: {
           txid: expect.any(String),
+          canBeMalleable: false,
         },
       });
     });
@@ -710,6 +712,7 @@ describe('KeyringRequestHandler', () => {
         pending: false,
         result: {
           txid: expect.any(String),
+          canBeMalleable: false,
         },
       });
     });